feat: website-responsive-002 - Keep dashboard header compact and usable on mobile
