+++ ui/src/features/factory-graph-editor/public/index.ts
@@ -1,0 +1,1 @@
+export { useFactoryGraphTouchPanePan } from "../hooks/selection/use-factory-graph-touch-pane-pan";
