+++ ui/src/features/header/components/tick-slider-control.test.tsx
@@ -132,6 +132,7 @@ describe("TickSliderControl", () => {
     expect(sliderShell?.className).toContain("gap-1.5");
     expect(sliderShell?.className).toContain("px-1");
     expect(sliderShell?.className).toContain("py-1");
+    expect(slider.className).toContain("h-10");
     expect(sliderShell?.className).not.toContain("border-outline");
     expect(sliderShell?.className).not.toContain("bg-surface-container-low");
     expect(sliderShell?.className).not.toContain("rounded-lg");
